LSL : stdint + ADC v1

+++ include/lsl_utils.h
@@ -1,6 +1,8 @@
 #ifndef __LSL_UTILS_H
 #define __LSL_UTILS_H
 
-void LSL_UTILS_DelayMs(unsigned int d);
+#include <stdint.h>
+
+void LSL_UTILS_DelayMs(uint16_t d);
 
 #endif // __LSL_UTILS_H
